<Feature, BugFix> Map, Gear
Weapons can be Infused with essence to recover their durability.
Armour and Accessories can no longer be repaired or inscribed.
Armour plates break when they reach minimum health.
Quick travel works correctly.
Region is automatically entered on arrival.
Fixed issue where items would not move inventory.
Only humans drop salt.
Food and Water sources can lose and regain their contents.
New temple region added, needs visual elements added.
Updated route fitter to be smoother, still needs to fit centre of screen
so as to not overlap with the ui.
Removed dotted paths along routes, now displays a flowing series of
trails.
Refactored some map generation files into one.
Removed player vision, now automatically shows neighbors to visited
nodes.
Basic region information is displayed in the region select menu.
Reduced region select size and removed region type information.
Weather food, water, and visibility now affects regions.
More balancing.

+++ Night Unity Files/Assets/Scripts/Game/Exploration/Weather/WeatherManager.cs
@@ -40,9 +40,9 @@ namespace Game.Exploration.Weather
                 string name = weatherNode.SelectSingleNode("Name").InnerText;
                 string type = weatherNode.SelectSingleNode("Type").InnerText;
                 int temperature = int.Parse(weatherNode.SelectSingleNode("Temperature").InnerText);
-                int visibility = int.Parse(weatherNode.SelectSingleNode("Visibility").InnerText);
-                int water = int.Parse(weatherNode.SelectSingleNode("Water").InnerText);
-                int food = int.Parse(weatherNode.SelectSingleNode("Food").InnerText);
+                float visibility = float.Parse(weatherNode.SelectSingleNode("Visibility").InnerText);
+                float water = float.Parse(weatherNode.SelectSingleNode("Water").InnerText);
+                float food = float.Parse(weatherNode.SelectSingleNode("Food").InnerText);
                 int duration = int.Parse(weatherNode.SelectSingleNode("Duration").InnerText);
                 XmlNode particleNode = weatherNode.SelectSingleNode("Particles");
                 float rain = float.Parse(particleNode.SelectSingleNode("Rain").InnerText);
